pyproject.toml
[build-system]
requires = [
    "setuptools>=68",
]
build-backend = "setuptools.build_meta"

[project]
name = 'rocm-docs-spellcheck'
version = "0.1.0"
authors=[
  {name="Gergely Meszaros", email="gergely@streamhpc.com"}
]
description ='Spell checking for ROCm documentation'
readme="README.md"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]
dependencies = [
  'importlib_metadata>=6.8;python_version<"3.10"',
  'importlib_resources>=6.1;python_version<"3.9"',
  "pyspelling>=2.9.0",
  "pyyaml>=6.0",
  "rich>=13.5",
  "typer>=0.9",
  "typing_extensions>=4.8",
]
requires-python = ">=3.8"

[project.urls]
repository="https://github.com/StreamHPC/rocm-docs-spellcheck"
documentation="https://docs.amd.com"

[project.optional-dependencies]
development = [
  "black>=22.1",
  "build>=0.10.0",
  "commitizen>=2.42",
  "isort>=5.12.0",
  "mypy>=1.3.0",
  "pip-tools>=6.13.0",
  "pre-commit>=3.3.2",
  "pytest>=7.4.2",
  "ruff>=0.0.269",
]

[project.scripts]
rocm-docs-spellcheck="rocm_docs_spellcheck:app"

[tool.setuptools.packages.find]
where=["src"]

[tool.setuptools.package-data]
rocm_docs_spellcheck = ["py.typed"]
"rocm_docs_spellcheck.data" = ["*.yaml", "wordlist.txt"]

[tool.black]
target-version = ["py38"]
line-length = 80
color = true

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
version_files = [
  ".pre-commit-config.yaml:# cz_version",
  "pyproject.toml:^version",
  "README.md:rev:",
  "src/rocm_docs_spellcheck/__version__.py:__version__"
]
tag_format = "v$version"
annotated_tag = true
major_version_zero = true

[tool.isort]
# https://github.com/timothycrosley/isort/
py_version = "38"
line_length = 80

known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true

[tool.mypy]
allow_redefinition = false
check_untyped_defs = true
color_output = true
disallow_any_generics = true
disallow_incomplete_defs = true
implicit_reexport = false
pretty = true
python_version = 3.8
show_column_numbers = true
show_error_codes = true
show_error_context = true
show_traceback = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.ruff]
select = ["ARG","C4","D","E","F","FA","N","PT","RET","RUF","SIM","UP","W"]
ignore = ["E501", "D203", "D213", "D4"]
target-version = "py38"
line-length = 80

[tool.ruff.pyupgrade]
# Typer relies on type information at runtime
keep-runtime-typing = true

[tool.ruff.per-file-ignores]
"src/**/*" = ["PT"]
"tests/conftest.py" = ["F401", "F403"]
"tests/test_*.py" = ["D"]

[tool.pytest.ini_options]
addopts = ["--doctest-modules"]
filterwarnings = """
ignore::DeprecationWarning
default::DeprecationWarning:rocm_docs_spellcheck
default::PendingDeprecationWarning:rocm_docs_spellcheck
ignore::PendingDeprecationWarning
default::DeprecationWarning:tests
default::DeprecationWarning:rocm_docs_spellcheck
"""
